Move to target module; Restore backwards compat
